+++ database/migrations/2024_12_14_045209_create_blog_table.php
@@ -13,17 +13,19 @@ class CreateBlogTable extends Migration
      */
     public function up()
     {
-        Schema::create('blog', function (Blueprint $table) {
-            $table->id('blog_id');
-            $table->text('title')->nullable();
-            $table->text('category')->nullable();
-            $table->text('image')->nullable();
-            $table->text('short_description')->nullable();
-            $table->text('long_description')->nullable();
-            $table->text('related_blogs')->nullable();
-            $table->enum('status', [0, 1])->comment('0 = Not Published, 1 = Published');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('blog')) {
+            Schema::create('blog', function (Blueprint $table) {
+                $table->id('blog_id');
+                $table->text('title')->nullable();
+                $table->text('category')->nullable();
+                $table->text('image')->nullable();
+                $table->text('short_description')->nullable();
+                $table->text('long_description')->nullable();
+                $table->text('related_blogs')->nullable();
+                $table->enum('status', [0, 1])->comment('0 = Not Published, 1 = Published');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -33,6 +35,8 @@ class CreateBlogTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('blog');
+        if (Schema::hasTable('blog')) {
+            Schema::dropIfExists('blog');
+        }
     }
 }
